# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Powered LMS API",
    "description": "Backend for an AI-powered Learning Management System: video ingestion and transcription, retrieval over transcripts (RAG), assessment generation, automatic answer evaluation and learning reports.\n\n**Typical flow:** `POST /videos` \u2192 poll `GET /videos/{id}` until `indexed` \u2192 `POST /videos/{id}/progress` (\u2265 threshold) \u2192 `POST /assessments` \u2192 `POST /attempts` \u2192 `GET /attempts/{id}/report`.\n\nErrors use one envelope: `{\"error\": {\"code\", \"message\", \"request_id\", \"details\"}}`.",
    "version": "1.0.0"
  },
  "paths": {
    "/videos": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "List videos",
        "operationId": "list_videos_videos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VideoOut"
                  },
                  "type": "array",
                  "title": "Response List Videos Videos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "Register a video and queue it for processing",
        "description": "Provide **either** a multipart `file` upload **or** `sample_path` (a file name inside `sample_data/videos/`). Processing (transcription \u2192 chunking \u2192 topic labelling \u2192 embedding) runs in the background; poll `status_url` or `task_url`. Bodies over `MAX_UPLOAD_MB` are rejected with **413** as they stream in.",
        "operationId": "create_video_videos_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_video_videos_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{video_id}": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Video status and metadata",
        "operationId": "get_video_videos__video_id__get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{video_id}/chunks": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Transcript chunks with topics and timestamps",
        "operationId": "get_chunks_videos__video_id__chunks_get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChunkOut"
                  },
                  "title": "Response Get Chunks Videos  Video Id  Chunks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{video_id}/progress": {
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "Report learner watch progress",
        "description": "Simulates the signal a video player would send. Progress never decreases.",
        "operationId": "report_progress_videos__video_id__progress_post",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{video_id}/retrieve": {
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "Semantic search within one video",
        "operationId": "retrieve_videos__video_id__retrieve_post",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/assessments": {
      "post": {
        "tags": [
          "assessments"
        ],
        "summary": "Generate an assessment for a watched video",
        "description": "Returns **409** unless the video is indexed and the learner's progress is at or above `COMPLETION_THRESHOLD`. Generates a mix of MCQ, true/false and short-answer questions grounded in the video's transcript. Correct answers are not included in the response.\n\n**Synchronous LLM call:** typically 20-40 s (longer while Groq rate-limits and we back off). Clients and any reverse proxy in front of the API need a read timeout of >= 120 s.",
        "operationId": "create_assessment_assessments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/attempts": {
      "post": {
        "tags": [
          "attempts"
        ],
        "summary": "Submit answers; returns per-question evaluation and generates the report",
        "description": "MCQ and true/false are graded deterministically; short answers are graded by an LLM judge grounded in the reference answer, rubric and source transcript. Questions without a response score 0. The learning report is generated in the same transaction.\n\n**Synchronous LLM calls:** typically 5-20 s (longer while Groq rate-limits and we back off). Clients and any reverse proxy in front of the API need a read timeout of >= 120 s.",
        "operationId": "submit_attempt_attempts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttemptCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Background job status",
        "description": "States: PENDING, STARTED, RETRY, SUCCESS, FAILURE (UNKNOWN for an id this process never saw\nin local mode). The video's own `status` field is the durable record.",
        "operationId": "get_task_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentCreate": {
        "properties": {
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "learner_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Learner Id",
            "examples": [
              "learner-001"
            ]
          },
          "num_questions": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Questions"
          },
          "mix": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "propertyNames": {
                  "enum": [
                    "mcq",
                    "true_false",
                    "short_answer"
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mix",
            "description": "Relative share per question type, e.g. {'mcq': 2, 'true_false': 1, 'short_answer': 2}"
          }
        },
        "type": "object",
        "required": [
          "video_id",
          "learner_id"
        ],
        "title": "AssessmentCreate"
      },
      "AssessmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "learner_id": {
            "type": "string",
            "title": "Learner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "video_id",
          "learner_id",
          "created_at",
          "questions"
        ],
        "title": "AssessmentOut"
      },
      "AttemptCreate": {
        "properties": {
          "assessment_id": {
            "type": "string",
            "title": "Assessment Id"
          },
          "learner_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Learner Id"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AnswerIn"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "learner_id",
          "answers"
        ],
        "title": "AttemptCreate"
      },
...
      "Body_create_video_videos_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "sample_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Path"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "title": "Body_create_video_videos_post"
      },
      "ChunkOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index"
          },
          "start_time": {
            "type": "number",
            "title": "Start Time"
          },
          "end_time": {
            "type": "number",
            "title": "End Time"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "chunk_index",
          "start_time",
          "end_time",
          "topic",
          "text"
        ],
        "title": "ChunkOut"
      },
...
      "ErrorOut": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorOut",
        "description": "The envelope every error response uses (4xx and 5xx)."
      },
...
      "ProgressIn": {
        "properties": {
          "learner_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Learner Id",
            "examples": [
              "learner-001"
            ]
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "description": "Fraction of the video watched (0-1)",
            "examples": [
              0.95
            ]
          }
        },
        "type": "object",
        "required": [
          "learner_id",
          "progress"
        ],
        "title": "ProgressIn"
      },
      "ProgressOut": {
        "properties": {
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "learner_id": {
            "type": "string",
            "title": "Learner Id"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "completion_threshold": {
            "type": "number",
            "title": "Completion Threshold"
          },
          "assessment_unlocked": {
            "type": "boolean",
            "title": "Assessment Unlocked"
          }
        },
        "type": "object",
        "required": [
          "video_id",
          "learner_id",
          "progress",
          "completion_threshold",
          "assessment_unlocked"
        ],
        "title": "ProgressOut"
      },
...
      "RetrieveIn": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Query",
            "examples": [
              "What is the base case of a recursive function?"
            ]
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RetrieveIn"
      },
      "RetrieveOut": {
        "properties": {
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RetrievedChunkOut"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "video_id",
          "query",
          "results"
        ],
        "title": "RetrieveOut"
      },
...
      "TaskOut": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "backend": {
            "type": "string",
            "title": "Backend"
          },
          "result": {
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "state",
          "backend"
        ],
        "title": "TaskOut"
      },
      "VideoAccepted": {
        "properties": {
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url"
          },
          "task_url": {
            "type": "string",
            "title": "Task Url"
          }
        },
        "type": "object",
        "required": [
          "video_id",
          "task_id",
          "status",
          "status_url",
          "task_url"
        ],
        "title": "VideoAccepted"
      },
      "VideoOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topics"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "error",
          "duration_s",
          "language",
          "chunk_count",
          "topics",
          "created_at",
          "updated_at"
        ],
        "title": "VideoOut"
      }
    }
  }
}